examen beheer rework

+++ app/Http/Controllers/Beheer/ExamenBeheerController.php
@@ -9,6 +9,7 @@ use Symfony\Component\HttpFoundation\Response;
 use App\Http\Controllers\Controller;
 use App\Models\Examen;
 use App\Models\ExamenMoment;
+use App\Models\Opleidingen;
 
 class ExamenBeheerController extends Controller
 {
@@ -35,8 +36,9 @@ class ExamenBeheerController extends Controller
     public function create()
     {
         abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
-
-        return view('beheer.examens.create');
+        $opleidingen = Opleidingen::all()->toArray();
+        //dd($opleidingen);
+        return view('beheer.examens.create', compact('opleidingen'));
     }
 
     /**
@@ -54,8 +56,8 @@ class ExamenBeheerController extends Controller
             'examen' => 'required',
             'crebo_nr' => 'required|integer|digits:5',
             'datum' => 'required',
-            'tijd' => 'required',
-            'plaatsen' => 'required|integer',
+            // 'tijd' => 'required',
+            // 'plaatsen' => 'required|integer',
             'geplande_docenten' => 'required',
             'examen_opgeven_begin' => 'required',
             'examen_opgeven_eind' => 'required',
@@ -64,17 +66,17 @@ class ExamenBeheerController extends Controller
         $examen->vak = $request->vak;
         $examen->examen = $request->examen;
         $examen->crebo_nr = $request->crebo_nr;
-        $examen->plaatsen = $request->plaatsen;
         $examen->geplande_docenten = $request->geplande_docenten;
         $examen->examen_opgeven_begin = $request->examen_opgeven_begin;
         $examen->examen_opgeven_eind = $request->examen_opgeven_eind;
         $examen->uitleg = $request->uitleg;
+        $examen->datum = $request->datum;
         $examen->save();
         
-        $moment->examenid = $examen->id;
-        $moment->datum = $request->datum;
-        $moment->tijd = $request->tijd;
-        $moment->save();
+        // $moment->examenid = $examen->id;
+        // $moment->tijd = $request->tijd;
+        // $moment->plaatsen = $request->plaatsen;
+        // $moment->save();
         
         return redirect()->route('examens.index')->with('success','Examen toegevoegd.');
     }
@@ -89,10 +91,12 @@ class ExamenBeheerController extends Controller
     {
         abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
+        $opleidingen = Opleidingen::all()->toArray();
+
         $examen = Examen::where('id', $id)->with( 'examen_moments')->get()->toArray();
         $examen = $examen[0];
 
-        return view('beheer.examens.show')->with(compact('examen'));
+        return view('beheer.examens.show')->with(compact('examen', 'opleidingen'));
     }
 
     /**
@@ -105,11 +109,13 @@ class ExamenBeheerController extends Controller
     {
         abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
+        $opleidingen = Opleidingen::all()->toArray();
+
         $examen = Examen::where('id', $id)->with( 'examen_moments')->get()->toArray();
         $examen = $examen[0];
         // dd($examen);
                  
-        return view('beheer.examens.edit')->with(compact('examen'));
+        return view('beheer.examens.edit')->with(compact('examen', 'opleidingen'));
         
     }
 
@@ -129,8 +135,8 @@ class ExamenBeheerController extends Controller
             'examen' => 'required',
             'crebo_nr' => 'required|integer|digits:5',
             'datum' => 'required',
-            'tijd' => 'required',
-            'plaatsen' => 'required|integer',
+            // 'tijd' => 'required',
+            // 'plaatsen' => 'required|integer',
             'geplande_docenten' => 'required',
             'examen_opgeven_begin' => 'required',
             'examen_opgeven_eind' => 'required',
@@ -141,19 +147,19 @@ class ExamenBeheerController extends Controller
         $examen->vak = is_null($request->vak) ? $examen->vak : $request->vak;
         $examen->examen = is_null($request->examen) ? $examen->examen : $request->examen;
         $examen->crebo_nr = is_null($request->crebo_nr) ? $examen->crebo_nr : $request->crebo_nr;
-        $examen->plaatsen = is_null($request->plaatsen) ? $examen->plaatsen : $request->plaatsen;
         $examen->geplande_docenten = is_null($request->geplande_docenten) ? $examen->geplande_docenten : $request->geplande_docenten;
         $examen->examen_opgeven_begin = is_null($request->examen_opgeven_begin) ? $examen->examen_opgeven_begin : $request->examen_opgeven_begin;
         $examen->examen_opgeven_eind = is_null($request->examen_opgeven_eind) ? $examen->examen_opgeven_eind : $request->examen_opgeven_eind;
         $examen->uitleg = is_null($request->uitleg) ? $examen->uitleg : $request->uitleg;
+        $examen->datum = is_null($request->datum) ? $examen->datum : $request->datum;
         $examen->save();
         
-        $moment = ExamenMoment::find($id);
-        $moment->datum = is_null($request->datum) ? $moment->datum : $request->datum;
-        $moment->tijd = is_null($request->tijd) ? $moment->tijd : $request->tijd;
-        $moment->save();
+        // $moment = ExamenMoment::find($id);
+        // $moment->tijd = is_null($request->tijd) ? $moment->tijd : $request->tijd;
+        // $moment->plaatsen = is_null($request->plaatsen) ? $moment->plaatsen : $request->plaatsen;
+        // $moment->save();
         
-            return redirect()->route('examens.index')->with('success','Examen aangepast.');
+            return redirect()->route('examens.show', $id)->with('success','Examen aangepast.');
         }else {
             return redirect()->route('examens.index')->with('error','Examen niet gevonden.');
         }
@@ -169,12 +175,15 @@ class ExamenBeheerController extends Controller
     {
         abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-        
-        $examen = Examen::find($id);
-        $examen->delete();
+        if (Examen::where('id', $id)->exists()) {
+            $examen = Examen::find($id);
+            $examen->delete();
         
 
         return redirect()->route('examens.index')->with('success','Examen verwijderd.');
+        }else {
+            return redirect()->route('examens.index')->with('error','Examen niet gevonden.');
+        }
     }
 
 
@@ -188,19 +197,63 @@ class ExamenBeheerController extends Controller
         return view('beheer.moments.create')->with(compact('examen'));
     }
 
-    public function examenMomentStore(Request $request, Examen $examen, ExamenMoment $moment)
+    public function examenMomentEdit(Request $request, $id)
+    {
+        abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        // $examen = Examen::where('id', $id)->get()->toArray();
+        // // dd($examen);
+        // $examen = $examen[0];
+        $moment = ExamenMoment::where('id', $id)->first()->toArray();
+        $examen = Examen::where( 'id', $moment['examenid'])->first()->toArray();
+     
+
+        return view('beheer.moments.edit')->with(compact('examen', 'moment'));
+    }
+
+
+    public function examenMomentDelete($id)
+    {
+        abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+
+    }
+
+    public function examenMomentStore(Request $request, Examen $examen, ExamenMoment $moment, $id)
     {
         abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $this->validate($request, [
-            'datum' => 'required',
             'tijd' => 'required',
+            'plaatsen' => 'required'
         ]);
         $moment->examenid = $request->id;
-        $moment->datum = $request->datum;
         $moment->tijd = $request->tijd;
+        $moment->plaatsen = $request->plaatsen;
         $moment->save();
 
-        return redirect()->route('examens.index')->with('success','Examen moment toegevoegd.');
+        return redirect()->route('examens.show', $id)->with('success','Examen moment toegevoegd.');
+    }
+
+    public function examenMomentUpdate(Request $request, $id)
+    {
+        abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        $this->validate($request, [
+            'tijd' => 'required',
+            'plaatsen' => 'required|integer',
+        ]);
+        
+    if (Examen::where('id', $id)->exists()) {
+        
+        $moment = ExamenMoment::find($id);
+        $moment->tijd = is_null($request->tijd) ? $moment->tijd : $request->tijd;
+        $moment->plaatsen = is_null($request->plaatsen) ? $moment->plaatsen : $request->plaatsen;
+        $moment->save();
+        
+            return redirect()->route('examens.show', $id)->with('success','Examen moment aangepast.');
+        }else {
+            return redirect()->route('examens.index')->with('error','Examen moment niet gevonden.');
+        }
     }
 }
